+++ src/motif.h
@@ -78,8 +78,8 @@ public:
 	void add_site(const int c, const int p, const bool s);
 	void clear_sites();
 	void remove_all_sites();
-	void calc_freq_matrix(int* fm) const;
-	void calc_freq_matrix(int* fm, const vector<float>& w) const;
+	void calc_freq_matrix(float* fm) const;
+	void calc_freq_matrix(float* fm, const vector<float>& w) const;
 	void freq_matrix_extended(vector<float>& fm) const;
 	void calc_score_matrix(double* sm) const;
 	void calc_score_matrix(double* sm, const vector<float>& w) const;
